backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
pydantic==2.10.3
pillow==11.0.0
openpyxl==3.1.5
python-multipart==0.0.19
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1